+++ react-components/src/components/FundingCredit/FundingCreditHowMuchFunding/FundingCreditHowMuchFunding.jsx
@@ -1,14 +1,17 @@
 import React, { memo, useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
+import Services from '@src/Services'
 import { Input } from '@src/components/Form/Input'
 import { useDebounce } from '@src/components/hooks/useDebounce'
 
 export const FundingCreditHowMuchFunding = memo(({ ...data }) => {
-  const { formFields, currency, value } = data
-  const [formValue, setFormValue] = useState(value)
+  const { formData, currency } = data
+  const [formValue, setFormValue] = useState(
+    formData['funding_amount_required']
+  )
 
-  const update = (field, value) => {
-    Services.updateFundingCreditOption({ ...field, ...value })
+  const update = (field) => {
+    Services.updateExportPlan(field)
       .then(() => {})
       .catch(() => {})
   }
@@ -16,31 +19,27 @@ export const FundingCreditHowMuchFunding = memo(({ ...data }) => {
   const debounceUpdate = useDebounce(update)
 
   const inputData = {
-    onChange: (e) => {
-      debugger
-      setFormValue(e['how_much_funding'])
-      debounceUpdate(e)
+    onChange: (fieldItem) => {
+      setFormValue(fieldItem['funding_amount_required'])
+      debounceUpdate({
+        funding_and_credit: fieldItem,
+      })
     },
     value: formValue,
     prepend: currency,
     hideLabel: true,
     label: 'How much funding you need',
-    id: 'how_much_funding',
+    id: 'funding_amount_required',
     placeholder: '0',
     tooltip: null,
     type: 'number',
-    field: 'how_much_funding',
+    field: 'funding_amount_required',
   }
 
-  return (
-    <>
-      <Input {...inputData} />
-    </>
-  )
+  return <Input {...inputData} />
 })
 
-FundingCreditHowMuchFunding.propTypes = {}
-
-// FundingCreditHowMuchFunding.defaultProps = {
-//   estimated_costs_per_unit: 0,
-// }
+FundingCreditHowMuchFunding.propTypes = {
+  formData: PropTypes.object,
+  currency: PropTypes.string.isRequired,
+}