+++ react-components/src/components/FundingCredit/FundingCreditTotalExportCost/FundingCreditTotalExportCost.jsx
@@ -1,14 +1,17 @@
 import React, { memo, useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
+import Services from '@src/Services'
 import { Input } from '@src/components/Form/Input'
 import { useDebounce } from '@src/components/hooks/useDebounce'
 
 export const FundingCreditTotalExportCost = memo(({ ...data }) => {
-  const { estimated_costs_per_unit, formFields, currency, value } = data
-  const [formValue, setFormValue] = useState(value)
+  const { estimated_costs_per_unit, formData, currency } = data
+  const [formValue, setFormValue] = useState(
+    formData['override_estimated_total_cost']
+  )
 
-  const update = (field, value) => {
-    Services.updateFundingCreditOption({ ...field, ...value })
+  const update = (field) => {
+    Services.updateExportPlan(field)
       .then(() => {})
       .catch(() => {})
   }
@@ -16,20 +19,21 @@ export const FundingCreditTotalExportCost = memo(({ ...data }) => {
   const debounceUpdate = useDebounce(update)
 
   const inputData = {
-    onChange: (e) => {
-      debugger
-      setFormValue(e['total_export_cost'])
-      debounceUpdate(e)
+    onChange: (fieldItem) => {
+      setFormValue(fieldItem['override_estimated_total_cost'])
+      debounceUpdate({
+        funding_and_credit: fieldItem,
+      })
     },
     value: formValue,
     prepend: currency,
     hideLabel: true,
     label: 'Total export cost',
-    id: 'total_export_cost',
+    id: 'override_estimated_total_cost',
     placeholder: '0',
     tooltip: null,
     type: 'number',
-    field: 'total_export_cost',
+    field: 'override_estimated_total_cost',
     example:
       estimated_costs_per_unit !== 0
         ? {
@@ -51,17 +55,6 @@ export const FundingCreditTotalExportCost = memo(({ ...data }) => {
 
   return (
     <>
-      <h3 className="h-s">Your total export cost</h3>
-      <p>
-        Your total export cost is how much it will cost your business to export
-        for one year.
-      </p>
-      <p className="m-b-0">To work this out you will need:</p>
-      <ul className="list-bullet m-t-xs">
-        <li>your total direct costs per unit</li>
-        <li>your total overhead costs</li>
-        <li>the number of units you want to export</li>
-      </ul>
       {estimated_costs_per_unit !== 0 ? (
         <p>
           To help you, we've created an estimate for you based on the figures
@@ -84,8 +77,6 @@ export const FundingCreditTotalExportCost = memo(({ ...data }) => {
 
 FundingCreditTotalExportCost.propTypes = {
   estimated_costs_per_unit: PropTypes.number.isRequired,
+  formData: PropTypes.object,
+  currency: PropTypes.string.isRequired,
 }
-
-// FundingCreditTotalExportCost.defaultProps = {
-//   estimated_costs_per_unit: 0,
-// }